requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.7
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
